Add files via upload
The creation of this group project was part of the 3-months cours as Java Developer at Generation Italy. 
We were in charge of creating a CMS for a clothing shop which an admin can use to create, update and delete products, categories and discount strategies.

I mainly had the responsibility to manage the login both back end and front end sides.